+++ terasoluna-gfw-common/pom.xml
@@ -63,14 +63,6 @@
         </dependency>
         <!-- == End Google == -->
 
-
-        <!-- == Begin Joda-Time == -->
-        <dependency>
-            <groupId>joda-time</groupId>
-            <artifactId>joda-time</artifactId>
-        </dependency>
-        <!-- == End Joda-Time == -->
-
         <!-- == Begin AspectJ == -->
         <dependency>
             <groupId>org.aspectj</groupId>
@@ -86,10 +78,6 @@
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.datatype</groupId>
-            <artifactId>jackson-datatype-joda</artifactId>
         </dependency>
         <!-- == End Jackson == -->
